plat: Update following introduction on `uklcpu`
The LCPU APIs, register context APIs, exception APIs have been moved
and namespaced under `uklcpu`. Therefore, update accordingly.

Checkpatch-Ignore: AVOID_EXTERNS
Checkpatch-Ignore: LONG_LINE_COMMENT
