+++ plat/common/arm/cache64.S
@@ -36,7 +36,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#include <uk/arch/lcpu.h>
+#include <uk/asm/lcpu.h>
 #include <uk/asm.h>
 
 /*
