+++ plat/common/arm/time.c
@@ -34,7 +34,8 @@
 #include <uk/ofw/fdt.h>
 #include <uk/assert.h>
 #include <uk/plat/time.h>
-#include <uk/plat/lcpu.h>
+#include <uk/asm/lcpu.h>
+#include <uk/lcpu.h>
 #include <uk/atomic.h>
 #include <uk/plat/common/cpu.h>
 #include <uk/plat/common/sections.h>
