+++ plat/kvm/arm/efi_post.c
@@ -6,7 +6,7 @@
 #include <kvm/efi.h>
 #include <uk/paging.h>
 #include <uk/plat/common/bootinfo.h>
-#include <uk/plat/lcpu.h>
+#include <uk/lcpu.h>
 #include <uk/plat/common/sections.h>
 
 void clean_and_invalidate_dcache_range(unsigned long, unsigned long);
@@ -24,7 +24,7 @@ static __u8 __align(16) uk_efi_bootstack[__PAGE_SIZE];
 
 void __noreturn uk_efi_jmp_to_kern(void)
 {
-	ukplat_lcpu_disable_irq();
+	uk_lcpu_disable_irq();
 
 	/* Invalidate the image from the data cache */
 	clean_and_invalidate_dcache_range(__BASE_ADDR, __END - __BASE_ADDR);
