+++ plat/kvm/arm/entry64.S
@@ -29,7 +29,8 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#include <uk/arch/lcpu.h>
+#include <uk/asm/lcpu.h>
+#include <uk/lcpu.h>
 #include <uk/arch/limits.h>
 #include <uk/asm.h>
 #include <uk/plat/common/sections.h>
