+++ plat/kvm/arm/pagetable64.S
@@ -32,7 +32,8 @@
  *
  * $FreeBSD$
  */
-#include <uk/arch/lcpu.h>
+#include <uk/asm/lcpu.h>
+#include <uk/lcpu.h>
 #include <uk/arch/limits.h>
 #include <uk/asm.h>
 #include <uk/reloc.h>
